requirements.txt

Flask-PyMongo==2.2.0
Flask-Session==0.3.1
gunicorn==19.9.0
Jinja2==2.10
requests==2.20.1
